- breaks apart PortfolioProject into 3 files - all portfolio card components now react-native friendly
